fixed bugs added questions and code refactored
